requirements.txt
APScheduler==3.7.0
beautifulsoup4==4.12.2
email-validator==2.0.0.post2
Flask==3.0.0
lxml==5.1.0
numpy==1.26.3
python-dotenv==1.0.0
requests==2.31.0
selenium==4.9.1
tldextract==5.1.1
webdriver-manager==4.0.0
whois==0.9.27
